refactor: split Redraw flag types
