+++ src/CADShared/Cad/Editor/Display/RedrawEx.cs
@@ -3,85 +3,6 @@
 namespace Fs.Fox.Cad;
 
 /// <summary>
-/// 亮显模式
-/// </summary>
-[Flags]
-public enum BrightEntity
-{
-    /// <summary>
-    /// 块更新
-    /// </summary>
-    RecordGraphicsModified = 1,
-
-    /// <summary>
-    /// 标注更新
-    /// </summary>
-    RecomputeDimensionBlock = 2,
-
-    /// <summary>
-    /// 重画
-    /// </summary>
-    Draw = 4,
-
-    /// <summary>
-    /// 亮显
-    /// </summary>
-    Highlight = 8,
-
-    /// <summary>
-    /// 亮显取消
-    /// </summary>
-    Unhighlight = 16,
-
-    /// <summary>
-    /// 显示图元
-    /// </summary>
-    VisibleTrue = 32,
-
-    /// <summary>
-    /// 隐藏图元
-    /// </summary>
-    VisibleFalse = 64,
-
-    /// <summary>
-    /// 平移更新,可以令ctrl+z撤回时候保证刷新
-    /// </summary>
-    MoveZero = 128,
-}
-
-/// <summary>
-/// 刷新模式
-/// </summary>
-[Flags]
-public enum BrightEditor
-{
-    /// <summary>
-    /// 刷新屏幕,图元不生成(例如块还是旧的显示)
-    /// </summary>
-    UpdateScreen = 1,
-
-    /// <summary>
-    /// 刷新全图
-    /// </summary>
-    Regen = 2,
-
-    /// <summary>
-    /// 清空选择集
-    /// </summary>
-    SelectionClean = 4,
-
-    /// <summary>
-    /// 视口外
-    /// </summary>
-    ViewportsFrom = 8,
-
-    /// <summary>
-    /// 视口内
-    /// </summary>
-    ViewportsIn = 16,
-}
-
-/// <summary>
 /// 重绘扩展
 /// </summary>
 public static class RedrawEx
